Ensured that the slideshow delay length is always a reasonable value
A user could change the Local Storage value in their browser settings, causing an overflow if the number is too big. The plugin now checks for both minimum and maximum values.

The delay is now stored in milliseconds, to make the comparisons easier. The parameter is also in milliseconds, to match the params of PhotoSwipe.